+++ History.md
@@ -3,6 +3,7 @@
 
 - support different ship to address
 - allow for diffent namespace prefixes #140
+- include exemption reason if doing intra community supply
 
 1.7.5
 =====